List of Untranslated Files
Was translating a few things, and realized the hardest part was finding which files needed to be translated still. So being the hacker I am I wrote a script to figure out what constitutes a non-Japanese character, and to then list all the files that have Japanese characters. I've provided the script, but it doesn't need to be re-run so much, as people can just edit needs-translation.txt when they finish translating a file, removing the line with that file's name. If the script still registers that file as Japanese, you can just add its name into the script manually. Such false positives should become vanishingly rare as more files are manually added as such.